Change transaction logic (use negative numbers) - save negative amount for expenses - Increase number of characters for amount to 16

+++ src/entities/Budget.ts
@@ -9,7 +9,7 @@ export class Budget {
 
   @Column({
     type: 'decimal',
-    precision: 10,
+    precision: 16,
     scale: 2,
   })
   limit: number;
